fix pattern proxy

+++ src/main/java/ru/aston/homework/modul5/proxy/BankAccount.java
@@ -1,7 +1,9 @@
 package ru.aston.homework.modul5.proxy;
 
+import java.math.BigDecimal;
+
 public interface BankAccount {
-    void deposit(double amount);
-    void withdraw(double amount);
-    double getBalance();
+    void deposit(BigDecimal amount);
+    void withdraw(BigDecimal amount);
+    BigDecimal getBalance();
 }
